Extract *View classes
